// src/swagger/model/adminToEditDto.ts
/**
 * PaymentAdvisoryPortalAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface AdminToEditDto { 
    userId?: number;
    empCode?: string;
    empName?: string;
    empRoleID?: number;
    active?: number;
    modifiedBy?: number;
    empEmail?: string;
    empMobNo?: string;
}